bilbo contribution!! Mummi contribution!! veriverse contribution!! LCTV contribution!! veriverse contribution!! LCTV great contribution!! catherinacortega, davidwisser, drmjg, frankly444, fuel, intelijens, vapaamies, linzeming, denvercowboy and qoooeo contribution!! Twitch.tv contribution!! Casey Muratori and Handmade group!! Beam.pro contribution!! J4Wx and jordanlee833 contribution!! kranalex contribution!! bidzapfc, intelijens, r00tkiller contribution!! catherina is back contribution!! Waiting for fuel contribution (bit anxious in fact...) !! Twitch contribution!! Beam.pro contribution!! Carol and Lucinho contribution!!
Thank you Carlos!!

Thank you Mummi!!

Carol contribution!!

at bilbo at veriverse!!
i7 6700 16GB RAM SSD OCZ 120GB SSD 240GB HD 2TB Windows 10 Pro

+++ appseed/aura/multithreading/service/service_plain_service.cpp
@@ -20,7 +20,7 @@ void plain_service::Stop(uint32_t ui)
 {
    System.m_bDoNotExitIfNoApplications = false;
    m_bStopping = true;
-   m_stopped.wait(millis(m_dwStopTimeout));
+   m_stopped.wait(millis((uint32_t) m_dwStopTimeout));
 }
 
 void plain_service::CallServiceThread()
